CLN: move pandas.tseries.tdi -> pandas.core.indexes.timedeltas
