+++ pandas/core/frame.py
@@ -84,7 +84,7 @@ from pandas.util.validators import validate_bool_kwarg
 
 from pandas.core.indexes.period import PeriodIndex
 from pandas.tseries.index import DatetimeIndex
-from pandas.tseries.tdi import TimedeltaIndex
+from pandas.core.indexes.timedeltas import TimedeltaIndex
 
 import pandas.core.base as base
 import pandas.core.common as com
