+++ pandas/core/indexes/period.py
@@ -25,7 +25,7 @@ from pandas.core.dtypes.generic import ABCSeries
 import pandas.tseries.frequencies as frequencies
 from pandas.tseries.frequencies import get_freq_code as _gfc
 from pandas.tseries.index import DatetimeIndex, Int64Index, Index
-from pandas.tseries.tdi import TimedeltaIndex
+from pandas.core.indexes.timedeltas import TimedeltaIndex
 from pandas.tseries.base import DatelikeOps, DatetimeIndexOpsMixin
 from pandas.tseries.tools import parse_time_string
 import pandas.tseries.offsets as offsets
