+++ pandas/core/indexes/timedeltas.py
