+++ pandas/core/resample.py
@@ -11,7 +11,7 @@ from pandas.core.groupby import (BinGrouper, Grouper, _GroupBy, GroupBy,
 
 from pandas.tseries.frequencies import to_offset, is_subperiod, is_superperiod
 from pandas.tseries.index import DatetimeIndex, date_range
-from pandas.tseries.tdi import TimedeltaIndex
+from pandas.core.indexes.timedeltas import TimedeltaIndex
 from pandas.tseries.offsets import DateOffset, Tick, Day, _delta_to_nanoseconds
 from pandas.core.indexes.period import PeriodIndex, period_range
 import pandas.core.common as com
