+++ pandas/plotting/_timeseries.py
@@ -8,7 +8,7 @@ from pandas.tseries.offsets import DateOffset
 import pandas.tseries.frequencies as frequencies
 from pandas.tseries.index import DatetimeIndex
 from pandas.core.indexes.period import PeriodIndex
-from pandas.tseries.tdi import TimedeltaIndex
+from pandas.core.indexes.timedeltas import TimedeltaIndex
 from pandas.io.formats.printing import pprint_thing
 import pandas.compat as compat
