+++ pandas/tests/series/test_timeseries.py
@@ -8,7 +8,7 @@ import pandas as pd
 import pandas.util.testing as tm
 from pandas._libs.tslib import iNaT
 from pandas.compat import lrange, StringIO, product
-from pandas.tseries.tdi import TimedeltaIndex
+from pandas.core.indexes.timedeltas import TimedeltaIndex
 from pandas.tseries.index import DatetimeIndex
 from pandas.tseries.offsets import BDay, BMonthEnd
 from pandas import (Index, Series, date_range, NaT, concat, DataFrame,
