+++ pandas/tseries/api.py
@@ -6,7 +6,7 @@
 
 from pandas.tseries.index import DatetimeIndex, date_range, bdate_range
 from pandas.tseries.frequencies import infer_freq
-from pandas.tseries.tdi import Timedelta, TimedeltaIndex, timedelta_range
+from pandas.core.indexes.timedeltas import Timedelta, TimedeltaIndex, timedelta_range
 from pandas.core.indexes.period import Period, PeriodIndex, period_range, pnow
 from pandas.tseries.timedeltas import to_timedelta
 from pandas._libs.lib import NaT
